cleaned up with correct int type

+++ src/ADC.h
@@ -6,8 +6,8 @@
 
 enum joystick_direction {LEFT, RIGHT, UP, DOWN, NEUTRAL};
 
-void ADC_read_all_channels(uint16_t* read_data);
-void ADC_print_all_channels(uint16_t* array);
+void ADC_read_all_channels(uint8_t* read_data);
+void ADC_print_all_channels(uint8_t* array);
 void ADC_select_channel(uint8_t channel);
